requirements.txt
deeplabcut==2.2.0.6
easydict==1.9
h5py==3.1.0
matplotlib==3.1.3
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.1.5
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.6.1
torch==1.8.0
torchvision==0.9.0
umap-learn==0.5.1
openpyxl

